Extract parameters of generalized hyperbolic distribution objects

The function coef returns the parameters of a generalized hyperbolic distribution object as a list. The user can choose between the “chi/psi”, the “alpha.bar” and the “alpha/delta” parametrization. The function coefficients is a synonym for coef.

# S4 method for ghyp
coef(object, type = c("chi.psi", "alpha.bar", "alpha.delta"))

# S4 method for ghyp
coefficients(object, type = c("chi.psi", "alpha.bar", "alpha.delta"))

Arguments

object
An object inheriting from class ghyp.

type
According to type the parameters of either the “chi/psi”, the “alpha.bar” or the “alpha/delta” parametrization will be returned. If type is missing, the parameters belonging to the parametrization of the construction are returned.

Value

If type is “chi.psi” a list with components:

lambda	Shape parameter.

chi	Shape parameter.

psi	Shape parameters.

mu	Location parameter.

sigma	Dispersion parameter.

gamma	Skewness parameter.

If type is “alpha.bar” a list with components:

lambda	Shape parameter.

alpha.bar	Shape parameter.

mu	Location parameter.

sigma	Dispersion parameter.

gamma	Skewness parameter.

If type is “alpha.delta” a list with components:

lambda	Shape parameter.

alpha	Shape parameter.

delta	Shape parameter.

mu	Location parameter.

Delta	Dispersion matrix with a determinant of 1 (only returned in the multivariate case).

beta	Shape and skewness parameter.

Details

Internally, the “chi/psi” parametrization is used. However, fitting is only possible in the “alpha.bar” parametrization as it provides the most convenient parameter constraints.

Note

A switch from either the “chi/psi” to the “alpha.bar” or from the “alpha/delta” to the “alpha.bar” parametrization is not yet possible.

See also

Examples

  ghyp.mv <- ghyp(lambda = 1, alpha.bar = 0.1, mu = rep(0,2), sigma = diag(rep(1,2)),
                  gamma = rep(0,2), data = matrix(rt(1000, df = 4), ncol = 2))
  ## Get parameters
  coef(ghyp.mv, type = "alpha.bar")
  coefficients(ghyp.mv, type = "chi.psi")

  ## Simple modification (do not modify slots directly e.g. object@mu <- 0:1)
  param <- coef(ghyp.mv, type = "alpha.bar")
  param$mu <- 0:1
  do.call("ghyp", param) # returns a new 'ghyp' object