Internal ghyp functions

Internal ghyp functions. These functions are not to be called by the user.

.abar2chipsi(alpha.bar, lambda, eps = .Machine$double.eps)

.besselM3(lambda = 9/2, x = 2, logvalue = FALSE)

.check.data(data, case = c("uv", "mv"), na.rm = TRUE,
            fit = TRUE, dim = NULL)

.check.gig.pars(lambda, chi, psi)

.check.norm.pars(mu, sigma, gamma, dimension)

.check.opt.pars(opt.pars, symmetric)

.fit.ghyp(object, llh = 0, n.iter = 0, converged = FALSE, error.code = 0,
          error.message = "", parameter.variance, fitted.params, aic,
          trace.pars = list())

.ghyp.model(lambda, chi, psi, gamma)

.t.transform(lambda)

.inv.t.transform(lambda.transf)

.integrate.moment.gig(x, moment = 1, ...)

.integrate.moment.ghypuv(x, moment = 1, ...)

.dghypuv(x, lambda = 1, chi = 1, psi = 1, alpha.bar = NULL,
         mu = 1, sigma = 1, gamma = 0, logvalue = FALSE)

.dghypmv(x, lambda, chi, psi, mu, sigma, gamma, logvalue = FALSE)

.mle.default(data, pdf, vars, opt.pars = rep(TRUE, length(vars)),
             transform = NULL, se = FALSE,
             na.rm = FALSE, silent = FALSE, ...)

.p.default(q, pdf, pdf.args, lower, upper, ...)

.q.default(p, pdf, pdf.args, interval, p.lower, ...)

.test.ghyp(object, case = c("ghyp", "univariate", "multivariate"))

.is.gaussian(object)

.is.univariate(object)

.is.symmetric(object)

.is.student.t(object, symmetric = NULL)

.get.stepAIC.ghyp(stepAIC.obj,
                  dist = c("ghyp", "hyp", "NIG", "VG", "t", "gauss"),
                  symmetric = FALSE)

Details

.abar2chipsi
Convert “alpha.bar” to “chi” and “psi” when using the “alpha.bar” parametrization.

.besselM3
Wrapper function for besselK.

.check.data
This function checks data for consistency. Only data objects of typ data.frame, matrix or numeric are accepted.

.check.gig.pars
Some combinations of the GIG parameters are not allowed. This function checks whether this is the case or not.

.check.norm.pars
This function simply checks if the dimensions match.

.check.opt.pars
When calling the fitting routines (fit.ghypuv and fit.ghypmv) a named vector containing the parameters which should not be fitted can be passed. By default all parameters will be fitted.

.fit.ghyp
This function is called by the functions fit.ghypuv and fit.ghypmv to create objects of class mle.ghyp and mle.ghyp.

.ghyp.model
Check if the parameters denote a special case of the generalized hyperbolic distribution.

.t.transfrom
Transformation function used in fit.ghypuv for parameter nu belonging to the Student-t distribution.

.inv.t.transfrom
The inverse of t.transfrom.

.integrate.moment.gig
This function is used when computing the conditional expectation of a generalized inverse gaussian distribution.

.integrate.moment.ghypuv
This function is used when computing the conditional expectation of a univariate generalized hyperbolic distribution.

.dghypuv
This function is used during the fitting procedure. Use dghyp to compute the density of generalized hyperbolic distribution objects.

.dghypmv
This function is used during the fitting procedure. Use dghyp to compute the density of generalized hyperbolic distribution objects.

.mle.default
This function serves as a generic function for maximum likelihood estimation. It is for internal use only. See fit.ghypuv which wraps this function.

.p.default
A generic distribution function integrator given a density function. See pghyp for a wrapper of this function.

.q.default
A generic quantile function calculator given a density function. See qghyp for a wrapper of this function.

.test.ghyp
This function tests whether the object is of class ghyp and sometimes whether it is univariate or multivariate according to the argument case and states a corresponding error if not.

.is.gaussian
Tests whether the object is of a gaussian type.

.is.symmetric
Tests whether the object is symmetric.

.is.student.t
Tests whether the object describes a Student-t distribution.

.is.univariate
Tests whether the object is a univariate ghyp-distribution.

.get.stepAIC.ghyp
Returns a specific model from a list returned by stepAIC.ghyp