Histogram for univariate generalized hyperbolic distributions

The function hist computes a histogram of the given data values and the univariate generalized hyperbolic distribution.

# S4 method for ghyp
hist(x, data = ghyp.data(x), gaussian = TRUE,
     log.hist = F, ylim = NULL, ghyp.col = 1, ghyp.lwd = 1,
     ghyp.lty = "solid", col = 1, nclass = 30, plot.legend = TRUE,
     location = if (log.hist) "bottom" else "topright", legend.cex = 1, ...)

Arguments

x
Usually a fitted univariate generalized hyperbolic distribution of class mle.ghyp. Alternatively an object of class ghyp and a data vector.

data
An object coercible to a vector.

gaussian
If TRUE the probability density of the normal distribution is plotted as a reference.

log.hist
If TRUE the logarithm of the histogramm is plotted.

ylim
The “y” limits of the plot.

ghyp.col
The color of the density of the generalized hyperbolic distribution.

ghyp.lwd
The line width of the density of the generalized hyperbolic distribution.

ghyp.lty
The line type of the density of the generalized hyperbolic distribution.

col
The color of the histogramm.

nclass
A single number giving the number of cells for the histogramm.

plot.legend
If TRUE a legend is drawn.

location
The location of the legend. See legend for possible values.

legend.cex
The character expansion of the legend.

...
Arguments passed to plot and qqghyp.

Value

No value is returned.

See also

Examples

  data(smi.stocks)
  univariate.fit <- fit.ghypuv(data = smi.stocks[,"SMI"],
                               opt.pars = c(mu = FALSE, sigma = FALSE),
                               symmetric = TRUE)
  hist(univariate.fit)