Pairs plot for multivariate generalized hyperbolic distributions

This function is intended to be used as a graphical diagnostic tool for fitted multivariate generalized hyperbolic distributions. An array of graphics is created and qq-plots are drawn into the diagonal part of the graphics array. The upper part of the graphics matrix shows scatter plots whereas the lower part shows 2-dimensional histogramms.

# S4 method for ghyp
pairs(x, data = ghyp.data(x), main = "'ghyp' pairwise plot",
      nbins = 30, qq = TRUE, gaussian = TRUE,
      hist.col = c("white", topo.colors(100)),
      spline.points = 150, root.tol = .Machine$double.eps^0.5,
      rel.tol = root.tol, abs.tol = root.tol^1.5, ...)

Arguments

x
Usually a fitted multivariate generalized hyperbolic distribution of class mle.ghyp. Alternatively an object of class ghyp and a data matrix.

data
An object coercible to a matrix.

main
The title of the plot.

nbins
The number of bins of the 2-d histogram.

qq
If TRUE qq-plots are drawn.

gaussian
If TRUE qq-plots with the normal distribution are plotted.

hist.col
A vector of colors of the 2-d histgram.

spline.points
The number of support points when computing the quantiles used by the qq-plot. Passed to qqghyp.

root.tol
The tolerance of the quantiles. Passed to uniroot via qqghyp.

rel.tol
The tolerance of the quantiles. Passed to integrate via qqghyp.

abs.tol
The tolerance of the quantiles. Passed to integrate via qqghyp.

...
Arguments passed to plot and axis.

See also

Examples

  data(smi.stocks)
  fitted.smi.stocks <- fit.NIGmv(data = smi.stocks[1:200, ])
  pairs(fitted.smi.stocks)