Quantile-Quantile Plot

This function is intended to be used as a graphical diagnostic tool for fitted univariate generalized hyperbolic distributions. Optionally a qq-plot of the normal distribution can be added.

qqghyp(object, data = ghyp.data(object), gaussian = TRUE, line = TRUE,
       main = "Generalized Hyperbolic Q-Q Plot",
       xlab = "Theoretical Quantiles", ylab = "Sample Quantiles",
       ghyp.pch = 1, gauss.pch = 6, ghyp.lty = "solid",
       gauss.lty = "dashed", ghyp.col = "black", gauss.col = "black",
       plot.legend = TRUE, location = "topleft", legend.cex = 0.8,
       spline.points = 150, root.tol = .Machine$double.eps^0.5,
       rel.tol = root.tol, abs.tol = root.tol^1.5, add = FALSE, ...)

Arguments

object
Usually a fitted univariate generalized hyperbolic distribution of class mle.ghyp. Alternatively an object of class ghyp and a data vector.

data
An object coercible to a vector.

gaussian
If TRUE a qq-plot of the normal distribution is plotted as a reference.

line
If TRUE a line is fitted and drawn.

main
An overall title for the plot.

xlab
A title for the x axis.

ylab
A title for the y axis.

ghyp.pch
A plotting character, i.e., symbol to use for quantiles of the generalized hyperbolic distribution.

gauss.pch
A plotting character, i.e., symbol to use for quantiles of the normal distribution.

ghyp.lty
The line type of the fitted line to the quantiles of the generalized hyperbolic distribution.

gauss.lty
The line type of the fitted line to the quantiles of the normal distribution.

ghyp.col
A color of the quantiles of the generalized hyperbolic distribution.

gauss.col
A color of the quantiles of the normal distribution.

plot.legend
If TRUE a legend is drawn.

location
The location of the legend. See legend for possible values.

legend.cex
The character expansion of the legend.

spline.points
The number of support points when computing the quantiles. Passed to qghyp.

root.tol
The tolerance of the quantiles. Passed to uniroot.

rel.tol
The tolerance of the quantiles. Passed to integrate.

abs.tol
The tolerance of the quantiles. Passed to integrate.

add
If TRUE the points are added to an existing plot window. The legend argument then becomes deactivated.

...
Arguments passed to plot.

See also

Examples

  data(smi.stocks)

  smi <- fit.ghypuv(data = smi.stocks[, "Swiss.Re"])

  qqghyp(smi, spline.points = 100)

  qqghyp(fit.tuv(smi.stocks[, "Swiss.Re"], symmetric = TRUE),
         add = TRUE, ghyp.col = "red", line = FALSE)